fix(health): fix cr problem, optimize example
