+++ packages/tuya-panel-health-sdk/src/components/barPercent/index.tsx
@@ -1,38 +1,11 @@
-import React, { ReactNode, Component, SFC, useMemo } from 'react';
-import { View, StyleSheet, ViewStyle, StyleProp } from 'react-native';
-import { Utils } from 'tuya-panel-kit';
-
-const { convertX: cx } = Utils.RatioUtils;
-
-export interface BarListProps<T> {
-  dataSource: T[];
-  renderItem: (item: T, index: number) => ReactNode;
-  style?: StyleProp<ViewStyle>;
-}
-
-export interface ItemProps {
-  flexPercent: number;
-  backgroundColor?: string;
-}
+import React, { SFC, useMemo } from 'react';
+import { ViewStyle, StyleProp } from 'react-native';
+import { BarList } from './BarList';
 
 export const sum = (arr: number[]) => {
   if (arr.length === 0) return arr;
   return [...arr].reduce((acc, val) => acc + val, 0);
 };
-
-export const Item: SFC<ItemProps> = (props: ItemProps) => {
-  const { flexPercent = 1, backgroundColor = '#BF73DE' } = props;
-  return <View style={[styles.item, { backgroundColor, flex: flexPercent }]} />;
-};
-
-export class BarList<T> extends Component<BarListProps<T>> {
-  static Item = Item;
-  render() {
-    const { dataSource, renderItem, style } = this.props;
-    const items = dataSource.map((item: T, index: number) => renderItem(item, index));
-    return <View style={[styles.container, style]}>{items}</View>;
-  }
-}
 
 export const WAKE_COLOR = '#FFCF3A';
 export const LIGHT_COLOR = '#E5C7F2';
@@ -82,16 +55,5 @@ const BarPercent: SFC<BarPercentProps> = (props: BarPercentProps) => {
   };
   return <BarList dataSource={data} style={style} renderItem={renderItem} />;
 };
-
-const styles = StyleSheet.create({
-  container: {
-    flexDirection: 'row',
-    height: cx(24),
-    overflow: 'hidden',
-  },
-  item: {
-    height: '100%',
-  },
-});
 
 export default BarPercent;
