+++ packages/tuya-panel-health-sdk/src/components/progess/index.tsx
@@ -116,3 +116,5 @@ class Progess extends React.Component<ProgessProps> {
 }
 
 export default Progess;
+
+
