+++ packages/tuya-panel-health-sdk/src/components/scrollRuler/index.tsx
@@ -1,6 +1,7 @@
 import React, { FC, ReactNode, useEffect, useMemo, useRef } from 'react';
-import { View, StyleSheet, FlatList, ColorValue, ViewStyle } from 'react-native';
+import { View, FlatList, ColorValue, ViewStyle } from 'react-native';
 import { Utils, TYText } from 'tuya-panel-kit';
+import styles from './style';
 
 const { convertX: cx, convertY: cy, width } = Utils.RatioUtils;
 
@@ -161,56 +162,5 @@ ScrollRuler.defaultProps = {
   scaleColor: '#666',
   formatValue: v => v,
 };
-
-const styles = StyleSheet.create({
-  container: {
-    width,
-  },
-  item: {
-    alignSelf: 'flex-start',
-    marginRight: cx(7),
-    width: cx(1),
-  },
-  itemWrap: {
-    flexDirection: 'row',
-    width: cx(80),
-  },
-  pointer: {
-    borderBottomColor: '#666',
-    borderBottomWidth: 4,
-    borderLeftColor: 'transparent',
-    borderLeftWidth: 4,
-    borderRightColor: 'transparent',
-    borderRightWidth: 5,
-    borderTopColor: 'transparent',
-    borderTopWidth: 5,
-    height: 0,
-    width: 0,
-  },
-  pointerBox: {
-    alignItems: 'center',
-    justifyContent: 'center',
-    marginBottom: cx(8),
-  },
-  title: {
-    color: '#666',
-    fontSize: cx(16),
-    fontWeight: 'bold',
-    marginBottom: cy(10),
-    textAlign: 'center',
-  },
-  value: {
-    color: '#666',
-    fontSize: cx(48),
-    fontWeight: 'bold',
-    marginBottom: cy(12),
-    marginTop: cy(18),
-  },
-  valueWrap: {
-    alignItems: 'flex-end',
-    flexDirection: 'row',
-    justifyContent: 'center',
-  },
-});
 
 export default ScrollRuler;
