+++ packages/tuya-panel-health-sdk/src/utils/index.tsx
@@ -1,27 +1,6 @@
-import { Animated, Easing } from 'react-native';
+const sum = (arr: number[]) => {
+  if (arr.length === 0) return arr;
+  return [...arr].reduce((acc, val) => acc + val, 0);
+};
 
-const createAnimation = ({
-  value,
-  toValue,
-  duration,
-  delay,
-  easing,
-  useNativeDriver,
-  isInteraction,
-}) => {
-  return Animated.timing(value, {
-    toValue,
-    duration: duration || 400,
-    delay: delay || 0,
-    easing: easing || Easing.linear,
-    useNativeDriver: useNativeDriver || false,
-    isInteraction: isInteraction || false,
-  });
-};
-const getImageUrl = url => {
-  return typeof url === 'string' ? { uri: url } : url;
-};
-const leftValue = data => {
-  return data > 0 ? Math.round(data) : Math.ceil(data);
-};
-export { createAnimation, getImageUrl, leftValue };
+export { sum };
